Estimate update from data received
the robot now computes the correction value for every measurement it receives.
It puts them together cumputing the update terms and updating the estimates out.
However, as seen from the simulation either the matrix are wrong (dh matrix are complex) or the calculations on the update terms need to be checked.
One step shows that the estimate gets much too worse so it is definatly not a paper inaccuracy as dz was too small for those changes.
It could be however that the measurement variance is high (not sure as its the same values as in the paper)