fix: direction of measurement and exploding matrices
there was a problem when calculating the difference in direction as the angle were expressed in different references.
All angles are now in (0,2*pi) and the reference has been restored.
the matrices to be inversed must stay on certain conditions, the value on its trace must never be 0 and values that are too low can be put to 0 for computational complexity.

However, some robots are still lounched into space and the simulation breaks after some steps